requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
selenium==4.9.0